Cleaned up the video pulse codec
