+++ Tests/PlayMode/Utils/VideoPulseCodec.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8d0341f1305a54e608a89f8a2b68175f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
